usdeletion_of_additional_view_files
