--- comparing_template/register/Features_view.py
@@ -1,4 +1,0 @@
-from django.shortcuts import render
-
-def features(request):
-    return render(request, "main/features.html")
