--- comparing_template/register/about_us_view.py
@@ -1,4 +1,0 @@
-from django.shortcuts import render
-
-def about(request):
-    return render(request, "main/about_us.html")
