--- comparing_template/register/homepage_view.py
@@ -1,4 +1,0 @@
-from django.shortcuts import render
-
-def homepage(request):
-    return render(request, "main/home.html")
