+++ comparing_template/register/views.py
@@ -14,4 +14,9 @@ def register(request):
         form = RegisterForm()
 
     return render(request, "register/register.html", {"form": form})
-
+def about(request):
+    return render(request, "main/about_us.html")
+def features(request):
+    return render(request, "main/features.html")
+def homepage(request):
+    return render(request, "main/home.html")
